backend/requirements.txt
fastapi
uvicorn
websockets
python-dotenv
deepgram-sdk
elevenlabs
google-generativeai
tavily-python
openai
gunicorn